requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
imagehash>=4.3.1
Flask>=2.3.0
Werkzeug>=2.3.0
psutil>=5.9.0 